crud refractoring

+++ src/library/models.py
@@ -1,5 +1,5 @@
 from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy import Column, Integer, String, ForeignKey, Boolean
+from sqlalchemy import Column, Integer, String, ForeignKey, Boolean, SmallInteger
 from sqlalchemy.orm import relationship
 
 from src.auth.models import User
@@ -28,7 +28,7 @@ class Library(Base):
     id = Column(Integer, primary_key=True, unique=True, index=True, autoincrement=True)
     user_id = Column(UUID(as_uuid=True), ForeignKey(User.id), primary_key=True)
     book_id = Column(Integer, ForeignKey(Book.id), primary_key=True)
-    rating = Column(Integer, nullable=True)
+    rating = Column(SmallInteger, nullable=True)
     is_saved_to_profile = Column(Boolean)
 
     book = relationship("Book", back_populates="libraries")
